Move VisualizationConfigModel to separate file, move getStateQueryGridModel, getStateModelId to QueryGridModel.ts
