+++ packages/components/src/index.ts
@@ -40,7 +40,13 @@ import {
 } from './internal/components/base/ServerContext';
 import { naturalSort, naturalSortByProperty } from './public/sort';
 import { AssayDefinitionModel, AssayDomainTypes, AssayLink } from './internal/AssayDefinitionModel';
-import { IGridLoader, IGridResponse, QueryGridModel } from './internal/QueryGridModel';
+import {
+    getStateModelId,
+    getStateQueryGridModel,
+    IGridLoader,
+    IGridResponse,
+    QueryGridModel
+} from './internal/QueryGridModel';
 import {
     applyDevTools,
     capitalizeFirstChar,
@@ -116,7 +122,7 @@ import { ToggleButtons } from './internal/components/buttons/ToggleButtons';
 import { Cards } from './internal/components/base/Cards';
 import { Footer } from './internal/components/base/Footer';
 
-import { EditorModel, getStateModelId, getStateQueryGridModel, IDataViewInfo } from './internal/models';
+import { EditorModel, IDataViewInfo } from './internal/models';
 import {
     clearSelected,
     createQueryGridModelFilteredBySample,
